+++ deltafi-action-kit/src/main/java/org/deltafi/actionkit/action/filter/FilterResult.java
@@ -1,4 +1,4 @@
-/**
+/*
  *    DeltaFi - Data transformation and enrichment platform
  *
  *    Copyright 2021-2023 DeltaFi Contributors <deltafi@deltafi.org>
@@ -22,6 +22,7 @@ import lombok.Getter;
 import org.deltafi.actionkit.action.Result;
 import org.deltafi.actionkit.action.egress.EgressResultType;
 import org.deltafi.actionkit.action.format.FormatResultType;
+import org.deltafi.actionkit.action.join.JoinResultType;
 import org.deltafi.actionkit.action.load.LoadResultType;
 import org.deltafi.actionkit.action.transform.TransformResultType;
 import org.deltafi.actionkit.action.validate.ValidateResultType;
@@ -36,8 +37,8 @@ import org.jetbrains.annotations.NotNull;
  */
 @Getter
 @EqualsAndHashCode(callSuper = true)
-public class FilterResult extends Result<FilterResult> implements
-        EgressResultType, FormatResultType, LoadResultType, TransformResultType, ValidateResultType {
+public class FilterResult extends Result<FilterResult> implements EgressResultType, FormatResultType, JoinResultType,
+        LoadResultType, TransformResultType, ValidateResultType {
 
     private final String filteredCause;