+++ deltafi-action-kit/src/main/java/org/deltafi/actionkit/action/join/JoinAction.java
@@ -37,7 +37,7 @@ public abstract class JoinAction<P extends ActionParameters> extends Action<P> {
     }
 
     @Override
-    public final JoinResult executeAction(ActionInput actionInput) {
+    public final JoinResultType executeAction(ActionInput actionInput) {
         return join(actionInput.getDeltaFile(), actionInput.getJoinedDeltaFiles(), actionInput.getActionContext(),
                 convertToParams(actionInput.getActionParams()));
     }
@@ -48,13 +48,15 @@ public abstract class JoinAction<P extends ActionParameters> extends Action<P> {
      * @param joinedDeltaFiles the DeltaFiles to join from
      * @param context the context for this action execution
      * @param params the parameters that configure the behavior of this action execution
-     * @return a {@link JoinResult} object containing results for the action execution
+     * @return a {@link JoinResult} object containing results for the action execution, an
+     * {@link org.deltafi.actionkit.action.error.ErrorResult} if an error occurs, or a
+     * {@link org.deltafi.actionkit.action.filter.FilterResult} if the joined DeltaFile should be filtered
      */
-    protected abstract JoinResult join(DeltaFile deltaFile, List<DeltaFile> joinedDeltaFiles, ActionContext context,
+    protected abstract JoinResultType join(DeltaFile deltaFile, List<DeltaFile> joinedDeltaFiles, ActionContext context,
             P params);
 
     @Override
-    protected JoinResult execute(@NotNull DeltaFile deltaFile, @NotNull ActionContext context,
+    protected JoinResultType execute(@NotNull DeltaFile deltaFile, @NotNull ActionContext context,
             @NotNull ActionParameters params) {
         throw new UnsupportedOperationException("Join actions require a list of joined DeltaFiles");
     }