+++ deltafi-python/src/deltafi/actiontype.py
@@ -23,6 +23,7 @@ class ActionType(Enum):
     INGRESS = "ingress"
     TRANSFORM = "transform"
     LOAD = "load"
+    JOIN = "join"
     DOMAIN = "domain"
     ENRICH = "enrich"
     FORMAT = "format"